Github-Codes/TensorFlow-CNN-Dataset/Judging-a-Book-by-its-Cover/requirements.txt
appdirs==1.4.3
appnope==0.1.0
beautifulsoup4==4.5.3
bleach==2.0.0
cycler==0.10.0
decorator==4.0.11
EasyProcess==0.2.3
entrypoints==0.2.2
gym==0.8.1
html5lib==0.999999999
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==6.0.0
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.3.0
notebook==4.4.1
numpy==1.12.0
olefile==0.44
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
patsy==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.0.0
prompt-toolkit==1.0.13
protobuf==3.2.0
ptyprocess==0.5.1
pyglet==1.2.4
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2016.10
PyVCF==0.6.8
PyVirtualDisplay==0.2.1
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.13.0
scikit-learn==0.18.1
scipy==0.19.0
seaborn==0.7.1
selenium==3.3.3
simplegeneric==0.8.1
six==1.10.0
sklearn==0.0
statsmodels==0.8.0
tensorflow==1.0.1
terminado==0.6
testpath==0.3
tornado==4.4.2
tqdm==4.11.2
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5
Werkzeug==0.12
widgetsnbextension==2.0.0
